PDDA-475 Moving xhbCppStagingInboundRepo method back into its child class as before
