+++ src/main/java/uk/gov/hmcts/pdda/business/AbstractControllerBean.java
@@ -10,7 +10,6 @@ import uk.gov.hmcts.pdda.business.entities.xhbconfigprop.XhbConfigPropRepository
 import uk.gov.hmcts.pdda.business.entities.xhbcourt.XhbCourtRepository;
 import uk.gov.hmcts.pdda.business.entities.xhbcppformatting.XhbCppFormattingRepository;
 import uk.gov.hmcts.pdda.business.entities.xhbcpplist.XhbCppListRepository;
-import uk.gov.hmcts.pdda.business.entities.xhbcppstaginginbound.XhbCppStagingInboundRepository;
 import uk.gov.hmcts.pdda.business.entities.xhbformatting.XhbFormattingRepository;
 
 @SuppressWarnings("PMD.NullAssignment")
@@ -26,7 +25,6 @@ public class AbstractControllerBean {
     private XhbCppFormattingRepository xhbCppFormattingRepository;
     private XhbCppListRepository xhbCppListRepository;
     private XhbFormattingRepository xhbFormattingRepository;
-    private XhbCppStagingInboundRepository xhbCppStagingInboundRepository;
 
     // For unit tests.
     protected AbstractControllerBean(EntityManager entityManager,
@@ -48,7 +46,6 @@ public class AbstractControllerBean {
         xhbCourtRepository = null;
         xhbConfigPropRepository = null;
         xhbCppFormattingRepository = null;
-        xhbCppStagingInboundRepository = null;
     }
 
     protected AbstractControllerBean(EntityManager entityManager) {
@@ -126,18 +123,6 @@ public class AbstractControllerBean {
             xhbFormattingRepository = new XhbFormattingRepository(getEntityManager());
         }
         return xhbFormattingRepository;
-    }
-
-    /**
-     * Retrieves a reference to the xhbCppStagingInboundRepository.
-     * 
-     * @return XhbCppStagingInboundRepository
-     */
-    public XhbCppStagingInboundRepository getXhbCppStagingInboundRepository() {
-        if (xhbCppStagingInboundRepository == null || !isEntityManagerActive()) {
-            xhbCppStagingInboundRepository = new XhbCppStagingInboundRepository(getEntityManager());
-        }
-        return xhbCppStagingInboundRepository;
     }
     
     protected boolean isEntityManagerActive() {
