+++ src/main/java/uk/gov/hmcts/pdda/business/services/cppstaginginboundejb3/AbstractCppStagingInboundControllerBean.java
@@ -12,6 +12,7 @@ import uk.gov.hmcts.pdda.business.entities.xhbconfigprop.XhbConfigPropRepository
 import uk.gov.hmcts.pdda.business.entities.xhbcourt.XhbCourtDao;
 import uk.gov.hmcts.pdda.business.entities.xhbcourt.XhbCourtRepository;
 import uk.gov.hmcts.pdda.business.entities.xhbcppstaginginbound.XhbCppStagingInboundDao;
+import uk.gov.hmcts.pdda.business.entities.xhbcppstaginginbound.XhbCppStagingInboundRepository;
 import uk.gov.hmcts.pdda.business.services.validation.ValidationService;
 import uk.gov.hmcts.pdda.business.services.validation.sax.FileEntityResolver;
 import uk.gov.hmcts.pdda.business.services.validation.sax.SaxValidationService;
@@ -36,6 +37,7 @@ public class AbstractCppStagingInboundControllerBean extends AbstractControllerB
 
     private CppStagingInboundHelper cppStagingInboundHelper;
     private ValidationService validationService;
+    private XhbCppStagingInboundRepository xhbCppStagingInboundRepository;
     
     public AbstractCppStagingInboundControllerBean() {
         super();
@@ -52,6 +54,12 @@ public class AbstractCppStagingInboundControllerBean extends AbstractControllerB
         this.validationService = validationService;
     }
     
+    @Override
+    protected void clearRepositories() {
+        super.clearRepositories();
+        xhbCppStagingInboundRepository = null;
+    }
+
     /**
      * Get courtId from XHB_COURT using crest court id (aka court code).
      * 
@@ -157,5 +165,12 @@ public class AbstractCppStagingInboundControllerBean extends AbstractControllerB
             validationService = new SaxValidationService(new FileEntityResolver());
         }
         return validationService;
+    }
+    
+    public XhbCppStagingInboundRepository getXhbCppStagingInboundRepository() {
+        if (xhbCppStagingInboundRepository == null || !isEntityManagerActive()) {
+            xhbCppStagingInboundRepository = new XhbCppStagingInboundRepository(getEntityManager());
+        }
+        return xhbCppStagingInboundRepository;
     }
 }
